feat: support <driver/i2c_master.h> in i2cdev
