+++ examples/i2cdev/default/CMakeLists.txt
@@ -3,6 +3,7 @@
 cmake_minimum_required(VERSION 3.5)
 
 set(EXTRA_COMPONENT_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/../../../components)
+set(COMPONENTS main driver i2cdev)
 
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
 project(i2c_scanner)
